backend/src/requirements.txt
plaid-python>=11.0.0
boto3>=1.26.0
python-jose>=3.3.0
requests>=2.28.0 